+++ cli/src/compute_budget.rs
@@ -1,6 +1,6 @@
 use {
     solana_clap_utils::compute_budget::ComputeUnitLimit,
-    solana_compute_budget::compute_budget_processor::MAX_COMPUTE_UNIT_LIMIT,
+    solana_compute_budget::compute_budget_limits::MAX_COMPUTE_UNIT_LIMIT,
     solana_rpc_client::rpc_client::RpcClient,
     solana_rpc_client_api::config::RpcSimulateTransactionConfig,
     solana_sdk::{
