+++ runtime-transaction/src/lib.rs
@@ -1,5 +1,6 @@
 #![cfg_attr(RUSTC_WITH_SPECIALIZATION, feature(min_specialization))]
 #![allow(clippy::arithmetic_side_effects)]
 
+pub mod instructions_processor;
 pub mod runtime_transaction;
 pub mod transaction_meta;
