+++ runtime-transaction/src/runtime_transaction.rs
@@ -10,10 +10,11 @@
 //!    ALT, RuntimeTransaction<SanitizedMessage> transits into Dynamically Loaded state,
 //!    with its dynamic metadata loaded.
 use {
-    crate::transaction_meta::{DynamicMeta, StaticMeta, TransactionMeta},
-    solana_compute_budget::compute_budget_processor::{
-        process_compute_budget_instructions, ComputeBudgetLimits,
+    crate::{
+        instructions_processor::process_compute_budget_instructions,
+        transaction_meta::{DynamicMeta, StaticMeta, TransactionMeta},
     },
+    solana_compute_budget::compute_budget_limits::ComputeBudgetLimits,
     solana_sdk::{
         hash::Hash,
         message::{AddressLoader, SanitizedMessage, SanitizedVersionedMessage},
